+++ hal_lib/smbus.h
@@ -18,19 +18,19 @@ typedef struct {
 
 Bus *new_bus(I2C_HandleTypeDef *hi2c, UART_HandleTypeDef *huart);
 
-long read_byte(Bus *bus, uint8_t addr);
+uint8_t read_byte(Bus *bus, uint8_t addr);
 
 void write_byte(Bus *bus, uint8_t addr, uint8_t data);
 
-long read_byte_data(Bus *bus, uint8_t addr, char cmd);
+uint8_t read_byte_data(Bus *bus, uint8_t addr, char cmd);
 
 void write_byte_data(Bus *bus, uint8_t addr, char cmd, uint8_t data);
 
-long read_word_data(Bus *bus, uint8_t addr, char cmd);
+uint16_t read_word_data(Bus *bus, uint8_t addr, char cmd);
 
 void write_word_data(Bus *bus, uint8_t addr, char cmd, uint16_t data);
 
-long *read_block_data(Bus *bus, uint8_t addr, char cmd, uint8_t size);
+uint8_t *read_block_data(Bus *bus, uint8_t addr, char cmd, uint8_t size);
 
 void write_block_data(Bus *bus, uint8_t addr, char cmd, uint8_t *data, uint8_t size);
